+++ src/lib/providers/network-guard/index.tsx
@@ -10,6 +10,7 @@ import {
   useCodeStore,
   useContractStore,
   usePublicProjectStore,
+  useVerifyModuleTaskStore,
 } from "lib/providers/store";
 import { formatUserKey } from "lib/utils";
 
@@ -30,6 +31,8 @@ export const NetworkGuard = observer(({ children }: NetworkGuardProps) => {
   const { setCodeUserKey, isCodeUserKeyExist } = useCodeStore();
   const { setContractUserKey, isContractUserKeyExist } = useContractStore();
   const { setProjectUserKey, isProjectUserKeyExist } = usePublicProjectStore();
+  const { setVerifyModuleTaskUserKey, isVerifyModuleTaskUserKeyExist } =
+    useVerifyModuleTaskStore();
 
   useEffect(() => {
     if (isHydrated) {
@@ -38,6 +41,7 @@ export const NetworkGuard = observer(({ children }: NetworkGuardProps) => {
       setCodeUserKey(userKey);
       setContractUserKey(userKey);
       setProjectUserKey(userKey);
+      setVerifyModuleTaskUserKey(userKey);
     }
   }, [
     isHydrated,
@@ -46,6 +50,7 @@ export const NetworkGuard = observer(({ children }: NetworkGuardProps) => {
     setCodeUserKey,
     setContractUserKey,
     setProjectUserKey,
+    setVerifyModuleTaskUserKey,
   ]);
 
   if (isHydrated && !(currentChainId in chainConfigs))
@@ -56,7 +61,8 @@ export const NetworkGuard = observer(({ children }: NetworkGuardProps) => {
     !isAccountUserKeyExist() ||
     !isCodeUserKeyExist() ||
     !isContractUserKeyExist() ||
-    !isProjectUserKeyExist()
+    !isProjectUserKeyExist() ||
+    !isVerifyModuleTaskUserKeyExist()
   )
     return <LoadingOverlay />;